Refactor orchestration
