+++ pkg/orchestration/k8s/bee.go
@@ -1,4 +1,4 @@
-package bee
+package k8s
 
 import (
 	"bytes"
@@ -14,10 +14,11 @@ import (
 	"github.com/ethersphere/beekeeper/pkg/k8s/service"
 	"github.com/ethersphere/beekeeper/pkg/k8s/serviceaccount"
 	"github.com/ethersphere/beekeeper/pkg/k8s/statefulset"
+	ob "github.com/ethersphere/beekeeper/pkg/orchestration/bee"
 )
 
 // compile check whether client implements interface
-var _ k8s.Bee = (*Client)(nil)
+var _ ob.Bee = (*Client)(nil)
 
 // Client manages communication with the Kubernetes
 type Client struct {
@@ -37,7 +38,7 @@ func NewClient(k8s *k8s.Client) (c *Client) {
 }
 
 // Create creates Bee node in the cluster
-func (c *Client) Create(ctx context.Context, o k8s.CreateOptions) (err error) {
+func (c *Client) Create(ctx context.Context, o ob.CreateOptions) (err error) {
 	// bee configuration
 	var config bytes.Buffer
 	if err := template.Must(template.New("").Parse(configTemplate)).Execute(&config, o.Config); err != nil {
