+++ pkg/orchestration/k8s/helpers.go
@@ -1,4 +1,4 @@
-package bee
+package k8s
 
 import (
 	"strconv"
